[eos] Split observable database into two singletons
The flattened database of observable entries needs to be accessible when
creating an expression observable (`ExpressionKinematicReader`). This change
splits the list of observable sections from the flattened database.

Use `ObservableSections` for the sectioned database, e.g., when creating the
documentation.

Use `ObservableEntries` for the flattened databasem, e.g., when querying
a single observable entry.